--- Assets/Code/TurboDemo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d0c751336c5d4564db9bba282b0e7926
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
