--- Assets/Code/TurboDemo/MarketManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c748561b6eb9b6c4aabf2e2cd2c5ba2e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
